src/requirements.txt
beautifulsoup4==4.12.3
pandas==2.2.2
python-decouple==3.8
Requests==2.32.3
SQLAlchemy==1.4.52
